+++ src/harlequin_databricks/cli_options.py
@@ -40,8 +40,11 @@ password = TextOption(
 
 auth_type = SelectOption(
     name="auth-type",
-    description="Set to `databricks-oauth` if using OAuth user-to-machine (U2M) authentication",
-    choices=["databricks-oauth"],
+    description=(
+        "Set to `databricks-oauth` or `azure-oauth` if using OAuth user-to-machine (U2M) "
+        "authentication"
+    ),
+    choices=["databricks-oauth", "azure-oauth"],
 )
 
 skip_legacy_indexing = FlagOption(
@@ -54,6 +57,24 @@ skip_legacy_indexing = FlagOption(
     ),
 )
 
+client_id = TextOption(
+    name="client-id",
+    description=(
+        "Service principal's UUID or Application ID for OAuth machine-to-machine (M2M) "
+        "authentication. To use OAuth M2M you need to install `databricks-sdk` as an extra into "
+        "your python environment."
+    ),
+)
+
+client_secret = TextOption(
+    name="client-secret",
+    description=(
+        "Service principal's secrete for OAuth when using OAuth machine-to-machine (M2M) "
+        "authentication. To use OAuth M2M you need to install `databricks-sdk` as an extra into "
+        "your python environment."
+    ),
+)
+
 DATABRICKS_ADAPTER_OPTIONS = [
     server_hostname,
     http_path,
@@ -62,4 +83,6 @@ DATABRICKS_ADAPTER_OPTIONS = [
     password,
     auth_type,
     skip_legacy_indexing,
+    client_id,
+    client_secret,
 ]